+++ src/main/java/de/maxhenkel/voicechat/voice/client/GroupPlayerIconOrientation.java
@@ -1,0 +1,7 @@
+package de.maxhenkel.voicechat.voice.client;
+
+public enum GroupPlayerIconOrientation {
+
+    VERTICAL, HORIZONTAL
+
+}